+++ lib/components/fo_button/fo_button_component.dart
@@ -11,30 +11,30 @@ import 'fo_button_event.dart';
     changeDetection: ChangeDetectionStrategy.OnPush)
 class FoButtonComponent implements OnDestroy {
   @Input()
-  String label;
+  String? label;
 
   @Input()
 
   /// Button icon, don't use when the button has a label (use leading-/trailingIcon for that)
-  String icon;
+  String? icon;
 
   @Input()
 
   /// Only use when the button also has a label
-  String leadingIcon;
+  String? leadingIcon;
 
   @Input()
 
   /// Only use when the button also has a label
-  String trailingIcon;
+  String? trailingIcon;
 
   /// Set text/icon color dynamically, overrides any color set by mixins and attributes
   @Input()
-  String themeColor;
+  String? themeColor;
 
   /// Set background color dynamically, overrides any color set by mixins and attributes.
   @Input()
-  String themeBackgroundColor;
+  String? themeBackgroundColor;
 
   @Input()
   bool disabled = false;
@@ -51,7 +51,8 @@ class FoButtonComponent implements OnDestroy {
     }
   }
 
-  final StreamController _triggerController = StreamController<FoButtonEvent>();
+  final StreamController<FoButtonEvent> _triggerController =
+      StreamController<FoButtonEvent>();
 
   FoButtonComponent();