+++ lib/components/fo_dropdown_list/fo_dropdown_option.dart
@@ -1,29 +1,29 @@
 class FoDropdownOption implements FoDropdownOptionRenderable {
-  Object id;
-  String icon;
-  String label;
-  String secondaryLabel;
-  String info;
-  String image;
-  List<String> tags;
+  Object? id;
+  String? icon;
+  String label = '';
+  String? secondaryLabel;
+  String? info;
+  String? image;
+  List<String> tags = [];
 
   @override
-  String get renderIcon => icon;
+  String? get renderIcon => icon;
 
   @override
-  Object get renderId => id;
+  Object get renderId => id == null ? '' : id!;
 
   @override
-  String get renderInfo => info;
+  String? get renderInfo => info;
 
   @override
   String get renderLabel => label;
 
   @override
-  String get renderSecondaryLabel => secondaryLabel;
+  String? get renderSecondaryLabel => secondaryLabel;
 
   @override
-  String get renderImage => image;
+  String? get renderImage => image;
 
   @override
   List<String> get renderTags => tags;
@@ -32,11 +32,11 @@ class FoDropdownOption implements FoDropdownOptionRenderable {
 /// Objects implementing this class can be rendered by a DropdownList
 abstract class FoDropdownOptionRenderable {
   Object get renderId;
-  String get renderIcon;
-  String get renderImage;
+  String? get renderIcon;
+  String? get renderImage;
   String get renderLabel;
-  String get renderSecondaryLabel;
-  String get renderInfo;
+  String? get renderSecondaryLabel;
+  String? get renderInfo;
 
   /// Use tags to include hidden filter keywords
   List<String> get renderTags;