+++ lib/components/fo_dropdown_select/fo_dropdown_select_component.dart
@@ -24,10 +24,10 @@ import '../fo_label/fo_label_component.dart';
     changeDetection: ChangeDetectionStrategy.OnPush)
 class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
   @Input()
-  String label;
+  String? label;
 
   @Input()
-  String actionButtonLabel;
+  String? actionButtonLabel;
 
   @Input()
   bool disabled = false;
@@ -43,7 +43,7 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
 
   /// Max height of the dropdown, in pixels
   @Input()
-  int dropdownMaxHeight;
+  int? dropdownMaxHeight;
 
   /// Make sure options doesn't extend beyond the viewport
   @Input()
@@ -52,10 +52,10 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
   @Input()
   bool materialIcons = true;
 
-  Map<String, List<FoDropdownOptionRenderable>> _options;
+  Map<String, List<FoDropdownOptionRenderable>> _options = {};
 
-  bool _optionsChanged;
-  Object _selectedId;
+  bool? _optionsChanged;
+  Object? _selectedId;
 
   final ChangeDetectorRef _changeDetectorRef;
 
@@ -65,14 +65,13 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
   final StreamController<FoButtonEvent> _actionButtonController =
       StreamController<FoButtonEvent>();
   final dom.Element _host;
-
   bool dropdownVisible = false;
+  FoDropdownOptionRenderable? selectedOption;
 
-  FoDropdownOptionRenderable selectedOption;
   @Input()
   bool showSearch = false;
   @ViewChild('selector')
-  dom.Element selectorElement;
+  dom.Element? selectorElement;
 
   FoDropdownSelectComponent(this._host, this._changeDetectorRef);
 
@@ -80,8 +79,8 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
   Stream<FoButtonEvent> get actionButtonTrigger =>
       _actionButtonController.stream;
 
-  int get dropdownWidth =>
-      selectorElement.getBoundingClientRect().width.round();
+  int? get dropdownWidth =>
+      selectorElement?.getBoundingClientRect().width.round();
 
   Map<String, List<FoDropdownOptionRenderable>> get options => _options;
 
@@ -91,23 +90,20 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
     _optionsChanged = true;
   }
 
-  Object get selectedId => _selectedId;
+  Object? get selectedId => _selectedId;
 
   @Input('selectedId')
-  set selectedId(Object id) {
+  set selectedId(Object? id) {
     selectedOption = null;
     _selectedId = id;
 
-    if (options != null) {
-      for (final category in options.keys) {
-        if (options[category] != null) {
-          selectedOption = options[category]
-              .firstWhere((e) => e?.renderId == id, orElse: () => null);
-
-          if (selectedOption != null) {
-            return;
-          }
-        }
+    for (final category in options.keys) {
+      if (options[category] != null) {
+        try {
+          selectedOption =
+              options[category]!.firstWhere((e) => e.renderId == id);
+          return;
+        } finally {}
       }
     }
   }
@@ -115,22 +111,21 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
   @Output('selectedIdChange')
   Stream<Object> get selectedIdChange => _selectedIdController.stream;
 
-  String get noFocusShadow =>
+  String? get noFocusShadow =>
       _host.attributes.containsKey('noFocusShadow') ? '1' : null;
 
-  String get square => _host.attributes.containsKey('square') ? '1' : null;
+  String? get square => _host.attributes.containsKey('square') ? '1' : null;
 
   @override
   void ngAfterChanges() {
     if (_optionsChanged == true) {
       // Update selected option
       for (final category in options.keys) {
-        selectedOption = options[category].firstWhere(
-            (option) => option?.renderId == _selectedId,
-            orElse: () => null);
-        if (selectedOption != null) {
+        try {
+          selectedOption = options[category]!
+              .firstWhere((option) => option.renderId == _selectedId);
           return;
-        }
+        } finally {}
       }
     }
   }
@@ -148,10 +143,7 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
 
   void onClick(dom.Event e) {
     if (disabled != true &&
-        options?.values
-                ?.where((option) => option?.isNotEmpty == true)
-                ?.isNotEmpty ==
-            true) {
+        options.values.where((option) => option.isNotEmpty).isNotEmpty) {
       Future.delayed(Duration(milliseconds: 100)).then((_) {
         dropdownVisible = !dropdownVisible;
         _changeDetectorRef.markForCheck();
@@ -162,7 +154,7 @@ class FoDropdownSelectComponent implements AfterChanges, OnDestroy {
       ..stopPropagation();
   }
 
-  int get dropdownTopOffset => square == null ? null : -1;
+  int? get dropdownTopOffset => square == null ? null : -1;
 
   void onSelect(FoDropdownOptionRenderable event) {
     dropdownVisible = false;