+++ lib/components/fo_icon/fo_icon_component.dart
@@ -1,12 +1,11 @@
 import 'package:angular/angular.dart';
-import 'package:angular_components/material_icon/material_icon.dart';
 
 /// An icon component. Make sure to have imported the icomoon font and style into your project.
 @Component(
     selector: 'fo-icon',
     templateUrl: 'fo_icon_component.html',
     styleUrls: ['fo_icon_component.css'],
-    directives: [MaterialIconComponent, NgFor, NgIf, NgStyle],
+    directives: [NgFor, NgIf, NgStyle],
     changeDetection: ChangeDetectionStrategy.OnPush)
 class FoIconComponent implements AfterChanges {
   String get iconClass => 'icon-$icon';
@@ -15,7 +14,7 @@ class FoIconComponent implements AfterChanges {
   String size = '1em !important';
 
   @Input()
-  String icon;
+  String icon = '';
 
   /// If this flag is set, fo_icon uses material icon instead
   @Input()