+++ lib/components/fo_number_input/fo_number_input_component.dart
@@ -31,28 +31,28 @@ import '../../pipes/capitalize_pipe.dart';
 )
 class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
   final html.Element _host;
-  ChangeFunction<int> _onChange;
+  ChangeFunction<int>? _onChange;
   NgControl control;
-  int value;
-  StreamSubscription<html.KeyboardEvent> _keyUpListener;
-  StreamSubscription<html.MouseEvent> _mouseUpListener;
-  StreamSubscription<html.TouchEvent> _touchEndListener;
+  int? value;
+  StreamSubscription<html.KeyboardEvent>? _keyUpListener;
+  StreamSubscription<html.MouseEvent>? _mouseUpListener;
+  StreamSubscription<html.TouchEvent>? _touchEndListener;
   final ChangeDetectorRef _changeDetectorRef;
-  Timer autoAddTimer;
-  Timer addStepTimer;
+  Timer? autoAddTimer;
+  Timer? addStepTimer;
   bool popupVisible = false;
-  String tabIndex = '0';
-  int tabIndexNum = 0;
+  String? tabIndex = '0';
+  int? tabIndexNum = 0;
 
   final String msgEnterValue = Intl.message('enter value', name: 'enter_value');
 
-  String get square => _host.attributes.containsKey('square') ? '1' : null;
+  String? get square => _host.attributes.containsKey('square') ? '1' : null;
 
   @Input()
   bool disabled = false;
 
   @Input()
-  String label;
+  String? label;
 
   @Input()
   int max = 1000;
@@ -67,7 +67,7 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
   bool materialIcons = true;
 
   @Input()
-  Map<String, List<FoDropdownOption>> options;
+  Map<String, List<FoDropdownOption>> options = {};
 
   @Input()
   String trailingText = '';
@@ -78,7 +78,7 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
       this._changeDetectorRef,
       this._host) {
     try {
-      tabIndexNum = tabIndex == null ? null : int.parse(tabIndex);
+      tabIndexNum = tabIndex == null ? null : int.parse(tabIndex!);
     } on FormatException catch (e) {
       print(e.message);
     }
@@ -87,37 +87,38 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
     _touchEndListener = html.document.onTouchEnd.listen(onMouseUp);
     _keyUpListener = html.document.onKeyUp.listen(onMouseUp);
 
-    if (control != null) control.valueAccessor = this;
+    control.valueAccessor = this;
   }
 
   String get formattedValue => value == null ? '-' : value.toString();
-  String get tabIndexAdd => tabIndexNum == null ? null : '${tabIndexNum + 1}';
+  String? get tabIndexAdd => tabIndexNum == null ? null : '${tabIndexNum! + 1}';
 
-  String get tabIndexSub => tabIndexNum == null ? null : '${tabIndexNum - 1}';
+  String? get tabIndexSub => tabIndexNum == null ? null : '${tabIndexNum! - 1}';
 
   void add(int count) {
-    value ??= (count is double) ? 0.0 : 0;
+    var v = value == null ? 0 : value!;
 
-    if (value + count >= min && value + count <= max) {
-      value += count;
+    if (v + count >= min && v + count <= max) {
+      v += count;
       if (_onChange != null) {
-        _onChange(value);
+        _onChange!(v);
       }
+      value = v;
       _changeDetectorRef.markForCheck();
     }
   }
 
   @override
   void ngOnDestroy() {
-    _mouseUpListener.cancel();
-    _touchEndListener.cancel();
-    _keyUpListener.cancel();
+    _mouseUpListener?.cancel();
+    _touchEndListener?.cancel();
+    _keyUpListener?.cancel();
   }
 
   @override
   void onDisabledChanged(bool isDisabled) {}
 
-  void onInputBlur(String v) {
+  void onInputBlur(String? v) {
     if (v == null) {
       value = 0;
     } else {
@@ -134,15 +135,15 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
         value = newValue;
       } on FormatException catch (e) {
         print(e);
-
+        value ??= 0;
         // Reset the value
-        if (value > min) {
+        if (value! > min) {
           if (value == max) {
             value = min;
             Future.delayed(Duration(milliseconds: 0)).then((_) {
               value = max;
               if (_onChange != null) {
-                _onChange(value);
+                _onChange!(value!);
               }
             });
           } else
@@ -153,7 +154,7 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
             Future.delayed(Duration(milliseconds: 0)).then((_) {
               value = min;
               if (_onChange != null) {
-                _onChange(value);
+                _onChange!(value!);
               }
             });
           } else {
@@ -163,7 +164,7 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
       }
     }
     if (_onChange != null) {
-      _onChange(value);
+      _onChange!(value!);
     }
   }
 
@@ -205,7 +206,7 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
   void registerOnTouched(TouchFunction f) {}
 
   void setValueClamped(String v) {
-    if (v == null || v.isEmpty)
+    if (v.isEmpty)
       value = 0;
     else {
       try {
@@ -213,11 +214,11 @@ class FoNumberInputComponent implements OnDestroy, ControlValueAccessor<int> {
       } on FormatException {
         value ??= 0;
       }
-      value = math.min(max, math.max(min, value));
+      value = math.min(max, math.max(min, value!));
     }
 
     if (_onChange != null) {
-      _onChange(value);
+      _onChange!(value!);
     }
   }