+++ lib/pipes/capitalize_pipe.dart
@@ -1,10 +1,9 @@
-import 'package:angular/di.dart' show Pipe, PipeTransform;
+import 'package:angular/angular.dart';
 
 /// Makes the first letter of a string uppercase
 /// patrick --> Patrick
 @Pipe('capitalize')
-class CapitalizePipe implements PipeTransform {
-  String transform(String value) => value == null || value.isEmpty
-      ? value
-      : '${value[0].toUpperCase()}${value.substring(1)}';
+class CapitalizePipe {
+  String transform(String value) =>
+      value.isEmpty ? value : '${value[0].toUpperCase()}${value.substring(1)}';
 }